Move BinaryProvider out of src/: only used in test/
BinaryProvider is not meant to be loaded at runtime. It does lots of
things that aren't acceptable at runtime, which is why it goes through
the care of generating a `deps.jl` file that doesn't import it.

This commit takes BinaryProvider out of `src/CxxWrap.jl` so that code
using CxxWrap doesn't wind up with BinaryProvider in its runtime
dependencies.

Since it's only used in that one test/build.jl file anyway, this is
fine.