Remove some automatic conversion and comparison
- Remove methods that allow comparing an object of type `CxxRef{T}` with
  one of type `T`. Note that other `Ref` types in Julia don't provide such
  methods either.
- Remove `convert` methods from `SmartPointer` instances to to other types.

In both cases, the goal is to avoid invalidations. To give an idea of how
bad the removed methods were, consider the number of invalidations as
computed with this snippet in a fresh Julia session:

    using SnoopCompileCore
    invalidations = @snoopr using CxxWrap;
    length(invalidations)

Before this patch, we get 20819 invalidations with Julia 1.8.4, and
13863 with Julia master.

After this patch, this is down to just 2695 in 1.8.4, respectively
1480 in Julia master.

The main issue with this patch is that it may break code relying on
these convenience methods. However, the resulting failures should be
easy to fix, and the fixed code should work fine with older CxxWrap
versions. I would nevertheless suggest to mark this as a breaking change
by following the appropriate semver rules.