Replace convert method for StdString to avoid invalidations
The Julia stdlib provides the following convert method:

    convert(::Type{T}, s::T) where {T<:AbstractString} = s
    convert(::Type{T}, s::AbstractString) where {T<:AbstractString} = T(s)::T

Thus it suffices to provide a conversion constructor for StdStringAllocated.

This is better than a convert method, because the latter introduces
invalidations in other packages (specifically in packages which do *not*
use CxxWrap but which are loaded together with CxxWrap into the same
Julia session).